GODRIVER-1844 make unified test runner runnable from outside integration
